SignalHandler moved from timers_manager to widget part
