+++ pype/modules/timers_manager/widget_user_idle.py
@@ -154,3 +154,15 @@ class WidgetUserIdle(QtWidgets.QWidget):
 
     def showEvent(self, event):
         self.bool_is_showed = True
+
+
+class SignalHandler(QtCore.QObject):
+    signal_show_message = QtCore.Signal()
+    signal_change_label = QtCore.Signal()
+    signal_stop_timers = QtCore.Signal()
+
+    def __init__(self, cls):
+        super().__init__()
+        self.signal_show_message.connect(cls.show_message)
+        self.signal_change_label.connect(cls.change_label)
+        self.signal_stop_timers.connect(cls.stop_timers)
